+++ thrust/system/hip/detail/scan_by_key.h
@@ -50,6 +50,68 @@ namespace hip_rocprim
 namespace __scan_by_key
 {
     template <typename Derived,
+              typename KeysInputIt,
+              typename ValuesInputIt,
+              typename ValuesOutputIt,
+              typename BinaryFunction,
+              typename KeyCompareFunction>
+    THRUST_HIP_RUNTIME_FUNCTION auto invoke_inclusive_scan_by_key(
+        execution_policy<Derived>& policy,
+        void*                      temporary_storage,
+        size_t&                    storage_size,
+        const KeysInputIt          keys_input,
+        const ValuesInputIt        values_input,
+        const ValuesOutputIt       values_output,
+        const size_t               size,
+        const BinaryFunction       scan_op,
+        const KeyCompareFunction   key_compare_op,
+        const hipStream_t          stream,
+        bool debug_sync) -> std::enable_if_t<decltype(nondeterministic(policy))::value, hipError_t>
+    {
+        return rocprim::inclusive_scan_by_key(temporary_storage,
+                                              storage_size,
+                                              keys_input,
+                                              values_input,
+                                              values_output,
+                                              size,
+                                              scan_op,
+                                              key_compare_op,
+                                              stream,
+                                              debug_sync);
+    }
+
+    template <typename Derived,
+              typename KeysInputIt,
+              typename ValuesInputIt,
+              typename ValuesOutputIt,
+              typename BinaryFunction,
+              typename KeyCompareFunction>
+    THRUST_HIP_RUNTIME_FUNCTION auto invoke_inclusive_scan_by_key(
+        execution_policy<Derived>& policy,
+        void*                      temporary_storage,
+        size_t&                    storage_size,
+        const KeysInputIt          keys_input,
+        const ValuesInputIt        values_input,
+        const ValuesOutputIt       values_output,
+        const size_t               size,
+        const BinaryFunction       scan_op,
+        const KeyCompareFunction   key_compare_op,
+        const hipStream_t          stream,
+        bool debug_sync) -> std::enable_if_t<!decltype(nondeterministic(policy))::value, hipError_t>
+    {
+        return rocprim::deterministic_inclusive_scan_by_key(temporary_storage,
+                                                            storage_size,
+                                                            keys_input,
+                                                            values_input,
+                                                            values_output,
+                                                            size,
+                                                            scan_op,
+                                                            key_compare_op,
+                                                            stream,
+                                                            debug_sync);
+    }
+
+    template <typename Derived,
               typename KeysInputIterator,
               typename ValuesInputIterator,
               typename ValuesOutputIterator,
@@ -75,16 +137,17 @@ namespace __scan_by_key
             return value_result;
 
         // Determine temporary device storage requirements.
-        hip_rocprim::throw_on_error(rocprim::inclusive_scan_by_key(NULL,
-                                                                   storage_size,
-                                                                   key_first,
-                                                                   value_first,
-                                                                   value_result,
-                                                                   num_items,
-                                                                   scan_op,
-                                                                   key_compare_op,
-                                                                   stream,
-                                                                   debug_sync),
+        hip_rocprim::throw_on_error(invoke_inclusive_scan_by_key(policy,
+                                                                 NULL,
+                                                                 storage_size,
+                                                                 key_first,
+                                                                 value_first,
+                                                                 value_result,
+                                                                 num_items,
+                                                                 scan_op,
+                                                                 key_compare_op,
+                                                                 stream,
+                                                                 debug_sync),
                                     "scan_by_key failed on 1st step");
 
         // Allocate temporary storage.
@@ -93,16 +156,17 @@ namespace __scan_by_key
         void *ptr = static_cast<void*>(tmp.data().get());
 
         // Run scan.
-        hip_rocprim::throw_on_error(rocprim::inclusive_scan_by_key(ptr,
-                                                                   storage_size,
-                                                                   key_first,
-                                                                   value_first,
-                                                                   value_result,
-                                                                   num_items,
-                                                                   scan_op,
-                                                                   key_compare_op,
-                                                                   stream,
-                                                                   debug_sync),
+        hip_rocprim::throw_on_error(invoke_inclusive_scan_by_key(policy,
+                                                                 ptr,
+                                                                 storage_size,
+                                                                 key_first,
+                                                                 value_first,
+                                                                 value_result,
+                                                                 num_items,
+                                                                 scan_op,
+                                                                 key_compare_op,
+                                                                 stream,
+                                                                 debug_sync),
                                     "scan_by_key failed on 2nd step");
 
         hip_rocprim::throw_on_error(
@@ -110,6 +174,74 @@ namespace __scan_by_key
             "inclusive_scan_by_key: failed to synchronize"
         );
         return value_result + num_items;
+    }
+
+    template <typename Derived,
+              typename KeysInputIt,
+              typename ValuesInputIt,
+              typename ValuesOutputIt,
+              typename InitialValueType,
+              typename BinaryFunction,
+              typename KeyCompareFunction>
+    THRUST_HIP_RUNTIME_FUNCTION auto invoke_exclusive_scan_by_key(
+        execution_policy<Derived>& policy,
+        void*                      temporary_storage,
+        size_t&                    storage_size,
+        const KeysInputIt          keys_input,
+        const ValuesInputIt        values_input,
+        const ValuesOutputIt       values_output,
+        const InitialValueType     initial_value,
+        const size_t               size,
+        const BinaryFunction       scan_op,
+        const KeyCompareFunction   key_compare_op,
+        const hipStream_t          stream,
+        bool debug_sync) -> std::enable_if_t<decltype(nondeterministic(policy))::value, hipError_t>
+    {
+        return rocprim::exclusive_scan_by_key(temporary_storage,
+                                              storage_size,
+                                              keys_input,
+                                              values_input,
+                                              values_output,
+                                              initial_value,
+                                              size,
+                                              scan_op,
+                                              key_compare_op,
+                                              stream,
+                                              debug_sync);
+    }
+
+    template <typename Derived,
+              typename KeysInputIt,
+              typename ValuesInputIt,
+              typename ValuesOutputIt,
+              typename InitialValueType,
+              typename BinaryFunction,
+              typename KeyCompareFunction>
+    THRUST_HIP_RUNTIME_FUNCTION auto invoke_exclusive_scan_by_key(
+        execution_policy<Derived>& policy,
+        void*                      temporary_storage,
+        size_t&                    storage_size,
+        const KeysInputIt          keys_input,
+        const ValuesInputIt        values_input,
+        const ValuesOutputIt       values_output,
+        const InitialValueType     initial_value,
+        const size_t               size,
+        const BinaryFunction       scan_op,
+        const KeyCompareFunction   key_compare_op,
+        const hipStream_t          stream,
+        bool debug_sync) -> std::enable_if_t<!decltype(nondeterministic(policy))::value, hipError_t>
+    {
+        return rocprim::deterministic_exclusive_scan_by_key(temporary_storage,
+                                                            storage_size,
+                                                            keys_input,
+                                                            values_input,
+                                                            values_output,
+                                                            initial_value,
+                                                            size,
+                                                            scan_op,
+                                                            key_compare_op,
+                                                            stream,
+                                                            debug_sync);
     }
 
     template <typename Derived,
@@ -140,19 +272,19 @@ namespace __scan_by_key
             return value_result;
 
         // Determine temporary device storage requirements.
-        hip_rocprim::throw_on_error(rocprim::exclusive_scan_by_key(NULL,
-                                                                   storage_size,
-                                                                   key_first,
-                                                                   value_first,
-                                                                   value_result,
-                                                                   init,
-                                                                   num_items,
-                                                                   scan_op,
-                                                                   key_compare_op,
-                                                                   stream,
-                                                                   debug_sync),
+        hip_rocprim::throw_on_error(invoke_exclusive_scan_by_key(policy,
+                                                                 NULL,
+                                                                 storage_size,
+                                                                 key_first,
+                                                                 value_first,
+                                                                 value_result,
+                                                                 init,
+                                                                 num_items,
+                                                                 scan_op,
+                                                                 key_compare_op,
+                                                                 stream,
+                                                                 debug_sync),
                                     "scan_by_key failed on 1st step");
-
 
         // Allocate temporary storage.
         thrust::detail::temporary_array<thrust::detail::uint8_t, Derived>
@@ -160,17 +292,18 @@ namespace __scan_by_key
         void *ptr = static_cast<void*>(tmp.data().get());
 
         // Run scan.
-        hip_rocprim::throw_on_error(rocprim::exclusive_scan_by_key(ptr,
-                                                                   storage_size,
-                                                                   key_first,
-                                                                   value_first,
-                                                                   value_result,
-                                                                   init,
-                                                                   num_items,
-                                                                   scan_op,
-                                                                   key_compare_op,
-                                                                   stream,
-                                                                   debug_sync),
+        hip_rocprim::throw_on_error(invoke_exclusive_scan_by_key(policy,
+                                                                 ptr,
+                                                                 storage_size,
+                                                                 key_first,
+                                                                 value_first,
+                                                                 value_result,
+                                                                 init,
+                                                                 num_items,
+                                                                 scan_op,
+                                                                 key_compare_op,
+                                                                 stream,
+                                                                 debug_sync),
                                     "scan_by_key failed on 2nd step");
 
         hip_rocprim::throw_on_error(