+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Documentation for rocThrust available at
 * Merged changes from upstream CCCL/thrust 2.3.2
   * Only the NVIDIA backend uses `tuple` and `pair` types from libcu++, other backends continue to
     use the original Thrust implementations and hence do not require libcu++ (CCCL) as a dependency.
+* Added the `thrust::hip::par_det` execution policy to enable bitwise reproducibility on algorithms that are not bitwise reproducible by default.
 
 ### Changes
 
@@ -17,7 +18,7 @@ Documentation for rocThrust available at
 
 ### Fixes
 
-* Fixed the HIP backend not passing `TestCopyIfNonTrivial` from the upstream (thrust) test suite. 
+* Fixed the HIP backend not passing `TestCopyIfNonTrivial` from the upstream (thrust) test suite.
 
 ## (Unreleased) rocThrust 3.1.0 for ROCm 6.2